requirements.txt
python-telegram-bot==21.6
python-dotenv==1.0.1
SQLAlchemy==2.0.36
APScheduler==3.10.4
pytz==2024.2
python-dateutil==2.9.0.post0
